refactor: Simpler
Remove  msg and focused on why `eq` always prints the same value for *self and *other
which would imply they are always the same. But that is not the case the if/asserts
always work. That means the "println!" is lying.

Using cargo-asm and looking at the output I see why, `eq` is testing 2 pointers:

 $ cargo asm --rust --no-color "<&dyn trait_based_state_machine::State<trait_based_state_machine::SwitchSm> as core::cmp::PartialEq>::eq"
 file /rustc/a55dd71d5fb0ec5a6a3a9e8c27b2127ba491ce52/library/core/src/ptr/mod.rs does not exist!
  fn eq(&self, other: &Self) -> bool {
      xor     rdi, rdx
      xor     rsi, rcx
      or      rsi, rdi
      sete    al
  }
  ret

So it's testing both the pointer to  the "struct SwitchSm" which is always the same as
`struct SwitchSm` is zero length, but also the address of the "vtable" that follows it.
Of note, the call to ` process` is actually not indirect throught the vtable but is a direct
 call    <trait_based_state_machine::StateOff as trait_based_state_machine::State<trait_based_state_machine::SwitchSm>>::process

Here is an example output of `cargo asm`:
 $ cargo asm --rust --no-color "<&dyn trait_based_state_machine::State<trait_based_state_machine::SwitchSm> as core::cmp::PartialEq>::eq"
 file /rustc/a55dd71d5fb0ec5a6a3a9e8c27b2127ba491ce52/library/core/src/ptr/mod.rs does not exist!
  fn eq(&self, other: &Self) -> bool {
      xor     rdi, rdx
      xor     rsi, rcx
      or      rsi, rdi
      sete    al
  }
  ret