refactor: Into workspace with modules
The most important thing this shows is that `&StateOff` is in `main` is
different from the `&StateOff` in switchsm/src/lib.rs and thus using the
`is_state_on()` and `is_state_off()` is required.

Also, looking at the generated code testing if the light is "on/off" using
`current_state` is 2 64 bit compares and thus slower than having a
a specific field in StateMachine which is this case could be just a `bool`.

Added profile.release and profile.rele3ase-with-debug with lto and
codegen-units. This works but optimizes away the init creation of
`let switch` where we test against local &StateOff as it knows it's
"does nothing".